+++ android/src/main/kotlin/com/airship/flutter/AirshipPluginVersion.kt
@@ -2,6 +2,6 @@ package com.airship.flutter
 
 class AirshipPluginVersion {
     companion object {
-        const val AIRSHIP_PLUGIN_VERSION = "5.5.0"
+        const val AIRSHIP_PLUGIN_VERSION = "6.0.0"
     }
 }