+++ example/lib/screens/message_view.dart
@@ -2,17 +2,18 @@ import 'package:flutter/material.dart';
 import 'package:airship_flutter/airship_flutter.dart';
 import 'package:airship_example/styles.dart';
 import 'package:flutter/services.dart';
+import 'package:collection/collection.dart';
 
 class MessageView extends StatefulWidget {
   final String messageId;
 
-  MessageView({this.messageId});
+  MessageView({required this.messageId});
 
   @override
   _MessageViewState createState() => _MessageViewState();
 }
 
-class _MessageViewState extends State<MessageView>  {
+class _MessageViewState extends State<MessageView> {
   bool isLoading = true;
 
   @override
@@ -26,30 +27,31 @@ class _MessageViewState extends State<MessageView>  {
     return FutureBuilder<List<InboxMessage>>(
         future: Airship.inboxMessages,
         builder: (context, snapshot) {
-          InboxMessage message;
-
           List<InboxMessage> list = [];
 
           if (snapshot.hasData) {
-            list = List<InboxMessage>.from(snapshot.data);
+            list = snapshot.data!;
           }
 
-          message = list.firstWhere((thisMessage) =>
-          widget.messageId == thisMessage.messageId,
-              orElse: () => null);
+          InboxMessage? message = list.firstWhereOrNull(
+              (thisMessage) => widget.messageId == thisMessage.messageId);
 
           return Scaffold(
             appBar: AppBar(
               title: message != null ? Text("${message.title}") : Container(),
               backgroundColor: Styles.background,
             ),
-            body: Stack(
-                children: <Widget>[
-                  isLoading
-                      ? Center(child: CircularProgressIndicator())
-                      : Container(),
-                  InboxMessageView(messageId: widget.messageId, onLoadStarted: handleLoadStarted, onLoadFinished: handleLoadFinished, onLoadError: handleLoadError, onClose: handleClose)
-                ]),
+            body: Stack(children: <Widget>[
+              isLoading
+                  ? Center(child: CircularProgressIndicator())
+                  : Container(),
+              InboxMessageView(
+                  messageId: widget.messageId,
+                  onLoadStarted: handleLoadStarted,
+                  onLoadFinished: handleLoadFinished,
+                  onLoadError: handleLoadError,
+                  onClose: handleClose)
+            ]),
           );
         });
   }
@@ -78,10 +80,10 @@ class _MessageViewState extends State<MessageView>  {
       showDialog(
           context: context,
           builder: (context) => AlertDialog(
-            title: Text(e.message != null ? e.message : "Unable to load message"),
-            content: Text(e.details != null ? e.details : ""),
-          )
-      );
+                title: Text(
+                    e.message != null ? e.message! : "Unable to load message"),
+                content: Text(e.details != null ? e.details : ""),
+              ));
     });
   }