+++ example/lib/screens/named_user_add.dart
@@ -9,7 +9,8 @@ class NamedUserAdd extends StatefulWidget {
   NamedUserAdd({this.updateParent});
 
   @override
-  _NamedUserAddState createState() => _NamedUserAddState(updateParent:updateParent);
+  _NamedUserAddState createState() =>
+      _NamedUserAddState(updateParent: updateParent);
 }
 
 class _NamedUserAddState extends State<NamedUserAdd> {
@@ -30,7 +31,7 @@ class _NamedUserAddState extends State<NamedUserAdd> {
           title: Text("Add Named User"),
           backgroundColor: Styles.background,
         ),
-        body: FutureBuilder(
+        body: FutureBuilder<String?>(
           future: Airship.namedUser,
           builder: (context, snapshot) {
             return SafeArea(
@@ -40,8 +41,8 @@ class _NamedUserAddState extends State<NamedUserAdd> {
                   Padding(
                     padding: const EdgeInsets.all(12.0),
                     child: TextAddBar(
-                      label: snapshot.hasData ? snapshot.data : "Not set",
-                      onTap: (text){
+                      label: snapshot.hasData ? snapshot.data! : "Not set",
+                      onTap: (text) {
                         Airship.setNamedUser(text);
                         updateParent();
                         Navigator.pop(context);