+++ example/lib/widgets/notifications_enabled_button.dart
@@ -1,4 +1,3 @@
-import 'package:flutter/widgets.dart';
 import 'package:flutter/material.dart';
 import 'package:airship_example/styles.dart';
 
@@ -8,15 +7,17 @@ class NotificationsEnabledButton extends StatelessWidget {
   final onPressed;
 
   NotificationsEnabledButton({
-    @required this.onPressed,
+    required this.onPressed,
   });
 
   @override
   Widget build(BuildContext context) {
-    return Center (child: Padding(
+    return Center(
+        child: Padding(
             padding: const EdgeInsets.all(8.0),
             child: MaterialButton(
-                child: Text( "Enable Push",
+                child: Text(
+                  "Enable Push",
                   style: Styles.homeButtonText,
                 ),
                 color: Styles.airshipRed,
@@ -24,9 +25,6 @@ class NotificationsEnabledButton extends StatelessWidget {
                 height: 40,
                 minWidth: 400,
                 padding: EdgeInsets.symmetric(vertical: 35),
-                onPressed:onPressed
-            )
-        )
-    );
+                onPressed: onPressed)));
   }
 }