+++ lib/src/preference_center_config.dart
@@ -10,7 +10,6 @@ List<Map<String, dynamic>> _toList(dynamic json) {
 
 /// Preference center config object.
 class PreferenceCenterConfig {
-
   /// The ID of the preference center.
   final String identifier;
 
@@ -30,12 +29,16 @@ class PreferenceCenterConfig {
 
     try {
       var identifier = json["id"];
-      var display = json["display"] != null ? PreferenceCenterCommonDisplay._fromJson(json["display"]) : null;
-      var sections = PreferenceCenterSection._fromJsonList(_toList(json["sections"]));
+      var display = json["display"] != null
+          ? PreferenceCenterCommonDisplay._fromJson(json["display"])
+          : null;
+      var sections =
+          PreferenceCenterSection._fromJsonList(_toList(json["sections"]));
       return PreferenceCenterConfig._internal(identifier, display, sections);
     } catch (e) {
       print("Invalid config: $e");
     }
+    return null;
   }
 
   @override
@@ -45,48 +48,47 @@ class PreferenceCenterConfig {
 }
 
 /// Preference center condition type.
-enum PreferenceCenterConditionType {
-  notificationOptIn
-}
+enum PreferenceCenterConditionType { notificationOptIn }
 
 /// Preference center condition.
 class PreferenceCenterCondition {
-
   /// The condition type.
   final PreferenceCenterConditionType? type = null;
 
-  static List<PreferenceCenterCondition> _fromJsonList(List<Map<String, dynamic>> jsonList) {
-    return jsonList.map((item) => PreferenceCenterCondition._fromJson(item)).toList();
+  static List<PreferenceCenterCondition> _fromJsonList(
+      List<Map<String, dynamic>> jsonList) {
+    return jsonList
+        .map((item) => PreferenceCenterCondition._fromJson(item))
+        .toList();
   }
 
   static PreferenceCenterCondition _fromJson(Map<String, dynamic> json) {
-      var type = json["type"];
-      switch (type) {
-        case "notification_opt_in":
-          return PreferenceCenterNotificationOptInCondition._fromJson(json);
-      }
-      throw Exception("Invalid condition: " + type);
+    var type = json["type"];
+    switch (type) {
+      case "notification_opt_in":
+        return PreferenceCenterNotificationOptInCondition._fromJson(json);
+    }
+    throw Exception("Invalid condition: " + type);
   }
 }
 
 /// Preference center condition opt-in.
-enum PreferenceCenterConditionOptIn {
-  optIn,
-  optOut
-}
+enum PreferenceCenterConditionOptIn { optIn, optOut }
 
 /// Preference center notification opt-in condition.
-class PreferenceCenterNotificationOptInCondition implements PreferenceCenterCondition {
-
+class PreferenceCenterNotificationOptInCondition
+    implements PreferenceCenterCondition {
   /// The condition type.
-  final PreferenceCenterConditionType type = PreferenceCenterConditionType.notificationOptIn;
+  final PreferenceCenterConditionType type =
+      PreferenceCenterConditionType.notificationOptIn;
 
   /// The condition opt-in status.
   final PreferenceCenterConditionOptIn whenStatus;
 
   const PreferenceCenterNotificationOptInCondition._internal(this.whenStatus);
 
-  static PreferenceCenterNotificationOptInCondition _fromJson(Map<String, dynamic> json) {
+  static PreferenceCenterNotificationOptInCondition _fromJson(
+      Map<String, dynamic> json) {
     var whenStatus = PreferenceCenterConditionOptIn.optIn;
     switch (json["when_status"]) {
       case "opt_in":
@@ -110,7 +112,6 @@ class PreferenceCenterNotificationOptInCondition implements PreferenceCenterCond
 
 /// Preference center common display information.
 class PreferenceCenterCommonDisplay {
-
   /// The display title.
   final String? title;
 
@@ -133,7 +134,6 @@ class PreferenceCenterCommonDisplay {
 
 /// Preference center common display information with icon.
 class PreferenceCenterIconDisplay implements PreferenceCenterCommonDisplay {
-
   /// The display title.
   final String? title;
 
@@ -143,7 +143,8 @@ class PreferenceCenterIconDisplay implements PreferenceCenterCommonDisplay {
   /// The display icon.
   final String? icon;
 
-  const PreferenceCenterIconDisplay._internal(this.title, this.subtitle, this.icon);
+  const PreferenceCenterIconDisplay._internal(
+      this.title, this.subtitle, this.icon);
 
   static PreferenceCenterIconDisplay _fromJson(Map<String, dynamic> json) {
     var title = json["name"];
@@ -159,14 +160,10 @@ class PreferenceCenterIconDisplay implements PreferenceCenterCommonDisplay {
 }
 
 /// Preference center section type.
-enum PreferenceCenterSectionType {
-  common,
-  labeledSectionBreak
-}
+enum PreferenceCenterSectionType { common, labeledSectionBreak }
 
 /// Preference center section.
 abstract class PreferenceCenterSection {
-
   /// The section type.
   PreferenceCenterSectionType get type;
 
@@ -179,7 +176,8 @@ abstract class PreferenceCenterSection {
   /// A list of preference center conditions.
   List<PreferenceCenterCondition>? get conditions;
 
-  static List<PreferenceCenterSection> _fromJsonList(List<Map<String, dynamic>> jsonList) {
+  static List<PreferenceCenterSection> _fromJsonList(
+      List<Map<String, dynamic>> jsonList) {
     return jsonList.map((e) => _fromJson(e)).toList();
   }
 
@@ -198,7 +196,6 @@ abstract class PreferenceCenterSection {
 
 /// Preference center common section.
 class PreferenceCenterCommonSection implements PreferenceCenterSection {
-
   /// The section type.
   @override
   final PreferenceCenterSectionType type = PreferenceCenterSectionType.common;
@@ -215,12 +212,17 @@ class PreferenceCenterCommonSection implements PreferenceCenterSection {
   @override
   final List<PreferenceCenterCondition>? conditions;
 
-  const PreferenceCenterCommonSection._internal(this.display, this.items, this.conditions);
+  const PreferenceCenterCommonSection._internal(
+      this.display, this.items, this.conditions);
 
   static PreferenceCenterCommonSection _fromJson(Map<String, dynamic> json) {
-    var display = json["display"] != null ? PreferenceCenterCommonDisplay._fromJson(json["display"]) : null;
+    var display = json["display"] != null
+        ? PreferenceCenterCommonDisplay._fromJson(json["display"])
+        : null;
     var items = PreferenceCenterItem._fromJsonList(_toList(json["items"]));
-    var conditions = json["conditions"] != null ? PreferenceCenterCondition._fromJsonList(_toList(json["conditions"])) : null;
+    var conditions = json["conditions"] != null
+        ? PreferenceCenterCondition._fromJsonList(_toList(json["conditions"]))
+        : null;
     return PreferenceCenterCommonSection._internal(display, items, conditions);
   }
 
@@ -231,10 +233,10 @@ class PreferenceCenterCommonSection implements PreferenceCenterSection {
 }
 
 class PreferenceCenterLabeledSectionBreak implements PreferenceCenterSection {
-
   /// The section type.
   @override
-  final PreferenceCenterSectionType type = PreferenceCenterSectionType.labeledSectionBreak;
+  final PreferenceCenterSectionType type =
+      PreferenceCenterSectionType.labeledSectionBreak;
 
   /// The common display information.
   @override
@@ -248,11 +250,17 @@ class PreferenceCenterLabeledSectionBreak implements PreferenceCenterSection {
   @override
   final List<PreferenceCenterCondition>? conditions;
 
-  const PreferenceCenterLabeledSectionBreak._internal(this.display, this.conditions);
+  const PreferenceCenterLabeledSectionBreak._internal(
+      this.display, this.conditions);
 
-  static PreferenceCenterLabeledSectionBreak _fromJson(Map<String, dynamic> json) {
-    var display = json["display"] != null ? PreferenceCenterCommonDisplay._fromJson(json["display"]) : null;
-    var conditions = json["conditions"] != null ? PreferenceCenterCondition._fromJsonList(_toList(json["conditions"])) : null;
+  static PreferenceCenterLabeledSectionBreak _fromJson(
+      Map<String, dynamic> json) {
+    var display = json["display"] != null
+        ? PreferenceCenterCommonDisplay._fromJson(json["display"])
+        : null;
+    var conditions = json["conditions"] != null
+        ? PreferenceCenterCondition._fromJsonList(_toList(json["conditions"]))
+        : null;
     return PreferenceCenterLabeledSectionBreak._internal(display, conditions);
   }
 
@@ -272,7 +280,6 @@ enum PreferenceCenterItemType {
 
 /// Preference center item.
 abstract class PreferenceCenterItem {
-
   /// The item type.
   PreferenceCenterItemType get type;
 
@@ -282,7 +289,8 @@ abstract class PreferenceCenterItem {
   /// A list of preference center conditions.
   List<PreferenceCenterCondition>? get conditions;
 
-  static List<PreferenceCenterItem> _fromJsonList(List<Map<String, dynamic>> jsonList) {
+  static List<PreferenceCenterItem> _fromJsonList(
+      List<Map<String, dynamic>> jsonList) {
     return jsonList.map((e) => _fromJson(e)).toList();
   }
 
@@ -304,7 +312,6 @@ abstract class PreferenceCenterItem {
 
 /// Preference center alert item button.
 class PreferenceCenterAlertItemButton {
-
   /// The alert item button text.
   final String text;
 
@@ -314,14 +321,16 @@ class PreferenceCenterAlertItemButton {
   /// The alert item button actions.
   final Map<String, dynamic> actions;
 
-  const PreferenceCenterAlertItemButton._internal(this.text, this.contentDescription, this.actions);
+  const PreferenceCenterAlertItemButton._internal(
+      this.text, this.contentDescription, this.actions);
 
   static PreferenceCenterAlertItemButton _fromJson(Map<String, dynamic> json) {
     var text = json["text"];
     var contentDescription = json["content_description"];
     var actions = json["actions"];
 
-    return PreferenceCenterAlertItemButton._internal(text, contentDescription, actions);
+    return PreferenceCenterAlertItemButton._internal(
+        text, contentDescription, actions);
   }
 
   @override
@@ -332,7 +341,6 @@ class PreferenceCenterAlertItemButton {
 
 /// Preference center alert item.
 class PreferenceCenterAlertItem implements PreferenceCenterItem {
-
   /// The alert item type.
   @override
   final PreferenceCenterItemType type = PreferenceCenterItemType.alert;
@@ -348,12 +356,17 @@ class PreferenceCenterAlertItem implements PreferenceCenterItem {
   /// The alert item button.
   final PreferenceCenterAlertItemButton? button;
 
-  const PreferenceCenterAlertItem._internal(this.display, this.button, this.conditions);
+  const PreferenceCenterAlertItem._internal(
+      this.display, this.button, this.conditions);
 
   static PreferenceCenterAlertItem _fromJson(Map<String, dynamic> json) {
     var display = PreferenceCenterIconDisplay._fromJson(json["display"]);
-    var button = json["button"] != null ? PreferenceCenterAlertItemButton._fromJson(_toMap(json["button"])) : null;
-    var conditions = json["conditions"] != null ? PreferenceCenterCondition._fromJsonList(_toList(json["conditions"])) : null;
+    var button = json["button"] != null
+        ? PreferenceCenterAlertItemButton._fromJson(_toMap(json["button"]))
+        : null;
+    var conditions = json["conditions"] != null
+        ? PreferenceCenterCondition._fromJsonList(_toList(json["conditions"]))
+        : null;
     return PreferenceCenterAlertItem._internal(display, button, conditions);
   }
 
@@ -365,10 +378,10 @@ class PreferenceCenterAlertItem implements PreferenceCenterItem {
 
 /// Preference center channel subscription item.
 class PreferenceCenterChannelSubscriptionItem implements PreferenceCenterItem {
-
   /// The channel subscription item type.
   @override
-  final PreferenceCenterItemType type = PreferenceCenterItemType.channelSubscription;
+  final PreferenceCenterItemType type =
+      PreferenceCenterItemType.channelSubscription;
 
   /// The channel subscription item common display information.
   @override
@@ -381,13 +394,18 @@ class PreferenceCenterChannelSubscriptionItem implements PreferenceCenterItem {
   /// The subscription list id.
   final String subscriptionId;
 
-  const PreferenceCenterChannelSubscriptionItem._internal(this.display, this.subscriptionId, this.conditions);
+  const PreferenceCenterChannelSubscriptionItem._internal(
+      this.display, this.subscriptionId, this.conditions);
 
-  static PreferenceCenterChannelSubscriptionItem _fromJson(Map<String, dynamic> json) {
+  static PreferenceCenterChannelSubscriptionItem _fromJson(
+      Map<String, dynamic> json) {
     var display = PreferenceCenterIconDisplay._fromJson(json["display"]);
     var subscriptionId = json["subscription_id"];
-    var conditions = json["conditions"] != null ? PreferenceCenterCondition._fromJsonList(_toList(json["conditions"])) : null;
-    return PreferenceCenterChannelSubscriptionItem._internal(display, subscriptionId, conditions);
+    var conditions = json["conditions"] != null
+        ? PreferenceCenterCondition._fromJsonList(_toList(json["conditions"]))
+        : null;
+    return PreferenceCenterChannelSubscriptionItem._internal(
+        display, subscriptionId, conditions);
   }
 
   @override
@@ -398,10 +416,10 @@ class PreferenceCenterChannelSubscriptionItem implements PreferenceCenterItem {
 
 /// Preference center contact subscription item.
 class PreferenceCenterContactSubscriptionItem implements PreferenceCenterItem {
-
   /// The contact subscription item type.
   @override
-  final PreferenceCenterItemType type = PreferenceCenterItemType.contactSubscription;
+  final PreferenceCenterItemType type =
+      PreferenceCenterItemType.contactSubscription;
 
   /// The contact subscription item common display information.
   @override
@@ -414,13 +432,18 @@ class PreferenceCenterContactSubscriptionItem implements PreferenceCenterItem {
   /// The subscription list id.
   final String subscriptionId;
 
-  const PreferenceCenterContactSubscriptionItem._internal(this.display, this.subscriptionId, this.conditions);
+  const PreferenceCenterContactSubscriptionItem._internal(
+      this.display, this.subscriptionId, this.conditions);
 
-  static PreferenceCenterContactSubscriptionItem _fromJson(Map<String, dynamic> json) {
+  static PreferenceCenterContactSubscriptionItem _fromJson(
+      Map<String, dynamic> json) {
     var display = PreferenceCenterIconDisplay._fromJson(json["display"]);
     var subscriptionId = json["subscription_id"];
-    var conditions = json["conditions"] != null ? PreferenceCenterCondition._fromJsonList(_toList(json["conditions"])) : null;
-    return PreferenceCenterContactSubscriptionItem._internal(display, subscriptionId, conditions);
+    var conditions = json["conditions"] != null
+        ? PreferenceCenterCondition._fromJsonList(_toList(json["conditions"]))
+        : null;
+    return PreferenceCenterContactSubscriptionItem._internal(
+        display, subscriptionId, conditions);
   }
 
   @override
@@ -431,32 +454,41 @@ class PreferenceCenterContactSubscriptionItem implements PreferenceCenterItem {
 
 /// Preference center contact subscription group item component.
 class PreferenceCenterContactSubscriptionGroupItemComponent {
-
   /// The channel scopes.
   final List<ChannelScope> scopes;
 
   /// The contact subscription group item common display information.
   final PreferenceCenterCommonDisplay display;
 
-  const PreferenceCenterContactSubscriptionGroupItemComponent._internal(this.scopes, this.display);
+  const PreferenceCenterContactSubscriptionGroupItemComponent._internal(
+      this.scopes, this.display);
 
-  static List<PreferenceCenterContactSubscriptionGroupItemComponent> _fromJsonList(List<Map<String, dynamic>> jsonList) {
+  static List<PreferenceCenterContactSubscriptionGroupItemComponent>
+      _fromJsonList(List<Map<String, dynamic>> jsonList) {
     return jsonList.map((e) => _fromJson(e)).toList();
   }
 
-  static PreferenceCenterContactSubscriptionGroupItemComponent _fromJson(Map<String, dynamic> json) {
-    var scopes = List<String>.from(json["scopes"]).map((scopeString) => _parseScope(scopeString)).toList();
+  static PreferenceCenterContactSubscriptionGroupItemComponent _fromJson(
+      Map<String, dynamic> json) {
+    var scopes = List<String>.from(json["scopes"])
+        .map((scopeString) => _parseScope(scopeString))
+        .toList();
     var display = PreferenceCenterCommonDisplay._fromJson(json["display"]);
 
-    return PreferenceCenterContactSubscriptionGroupItemComponent._internal(scopes, display);
+    return PreferenceCenterContactSubscriptionGroupItemComponent._internal(
+        scopes, display);
   }
 
   static ChannelScope _parseScope(String scopeString) {
-    switch(scopeString.toLowerCase()) {
-      case "app": return ChannelScope.app;
-      case "web": return ChannelScope.web;
-      case "email": return ChannelScope.email;
-      case "sms": return ChannelScope.sms;
+    switch (scopeString.toLowerCase()) {
+      case "app":
+        return ChannelScope.app;
+      case "web":
+        return ChannelScope.web;
+      case "email":
+        return ChannelScope.email;
+      case "sms":
+        return ChannelScope.sms;
     }
     throw Exception("Invalid scope: $scopeString");
   }
@@ -468,11 +500,12 @@ class PreferenceCenterContactSubscriptionGroupItemComponent {
 }
 
 /// Preference center contact subscription group item.
-class PreferenceCenterContactSubscriptionGroupItem implements PreferenceCenterItem {
-
+class PreferenceCenterContactSubscriptionGroupItem
+    implements PreferenceCenterItem {
   /// The contact subscription group item type.
   @override
-  final PreferenceCenterItemType type = PreferenceCenterItemType.contactSubscriptionGroup;
+  final PreferenceCenterItemType type =
+      PreferenceCenterItemType.contactSubscriptionGroup;
 
   /// The contact subscription group item common display information.
   @override
@@ -488,16 +521,22 @@ class PreferenceCenterContactSubscriptionGroupItem implements PreferenceCenterIt
   /// A list of subscription group item component.
   final List<PreferenceCenterContactSubscriptionGroupItemComponent> components;
 
-  const PreferenceCenterContactSubscriptionGroupItem._internal(this.display, this.subscriptionId, this.conditions, this.components);
+  const PreferenceCenterContactSubscriptionGroupItem._internal(
+      this.display, this.subscriptionId, this.conditions, this.components);
 
-  static PreferenceCenterContactSubscriptionGroupItem _fromJson(Map<String, dynamic> json) {
+  static PreferenceCenterContactSubscriptionGroupItem _fromJson(
+      Map<String, dynamic> json) {
     var display = PreferenceCenterIconDisplay._fromJson(json["display"]);
     var subscriptionId = json["subscription_id"];
-    var conditions = json["conditions"] != null ? PreferenceCenterCondition._fromJsonList(_toList(json["conditions"])) : null;
+    var conditions = json["conditions"] != null
+        ? PreferenceCenterCondition._fromJsonList(_toList(json["conditions"]))
+        : null;
     var components =
-    PreferenceCenterContactSubscriptionGroupItemComponent._fromJsonList(_toList(json["components"]));
+        PreferenceCenterContactSubscriptionGroupItemComponent._fromJsonList(
+            _toList(json["components"]));
 
-    return PreferenceCenterContactSubscriptionGroupItem._internal(display, subscriptionId, conditions, components);
+    return PreferenceCenterContactSubscriptionGroupItem._internal(
+        display, subscriptionId, conditions, components);
   }
 
   @override